# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "contact": {
      "name": "Ploiu",
      "url": "https://github.com/ploiu"
    },
    "title": "Ploiu File Server",
    "description": "a self-hostable file server written in rust",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/password": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createPassword"
              }
            }
          }
        },
        "tags": [
          "api"
        ],
        "description": "sets the password to access the api. All non /api endpoints will not work unless a password is set",
        "responses": {
          "201": {
            "description": "password created"
          },
          "400": {
            "description": "password has already been set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "description": "error occurred setting password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          }
        }
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updatePassword"
              }
            }
          }
        },
        "tags": [
          "api"
        ],
        "description": "updates the password to access the api. All non /api endpoints will not work unless a password is set",
        "responses": {
          "204": {
            "description": "password updated"
          },
          "401": {
            "description": "Failed to set password. Check server logs for details"
          }
        }
      }
    },
...
    "/tags/{id}": {
      "parameters": [
        {
          "name": "id",
          "required": true,
          "schema": {
            "type": "number",
            "nullable": false
          },
          "in": "path"
        }
      ],
      "get": {
        "tags": [
          "tags"
        ],
        "description": "retrieves the tag with the passed id",
        "responses": {
          "200": {
            "description": "tag retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tagApi"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "no tag with the passed id was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      },
      "put": {
        "tags": [
          "tags"
        ],
        "description": "updates the tag with the passed id",
        "responses": {
          "200": {
            "description": "tag updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tagApi"
                }
              }
            }
          },
          "400": {
            "description": "a tag with the new name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "no tag with the passed id was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "tags"
        ],
        "description": "deletes the tag with the passed id. Does not indicate if an existing tag was deleted, only if there was an error",
        "responses": {
          "204": {
            "description": "tag deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      }
    },
    "/tags": {
      "post": {
        "tags": [
          "tags"
        ],
        "description": "creates a new tag with the passed name",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tagApi"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "tag successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tagApi"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      }
    }
  },
  "servers": [
    {
      "description": "localhost",
      "url": "https://localhost:8000"
    }
  ],
  "components": {
    "schemas": {
...
      "basicMessage": {
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "tagApi": {
        "properties": {
          "id": {
            "nullable": true,
            "type": "number",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "nullable": false
          }
        }
      },
      "createPassword": {
        "properties": {
          "username": {
            "type": "string",
            "nullable": false
          },
          "password": {
            "type": "string",
            "nullable": false
          }
        }
      },
      "updatePassword": {
        "properties": {
          "oldAuth": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "nullable": false
              },
              "password": {
                "type": "string",
                "nullable": false
              }
            }
          },
          "newAuth": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "nullable": false
              },
              "password": {
                "type": "string",
                "nullable": false
              }
            }
          }
        }
      }
    },
    "responses": {
      "unauthorizedResponse": {
        "description": "missing or invalid password",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/basicMessage"
            }
          }
        }
      },
      "genericErrorResponse": {
        "description": "something failed on the db or file system side of the operation. Check server logs.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/basicMessage"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}